refactor: directory structure optimization
