+++ scripts/apis/cnnamestr.mjs
