+++ src/utils/index.ts
@@ -21,6 +21,11 @@ import {
   getAllWoodWords,
   getAllWords,
 } from './dict';
+import {
+  pickRandomEle,
+  pickRandomSingleEle,
+  safePickSingleEleByAlgorithm,
+} from './random';
 import { getMaxSetSize } from './size';
 
 /**
@@ -243,30 +248,6 @@ export function handleuUniqueSrunamePartEdgeCase(
 
 /**
  * @private
- * 通过算法从数组中随机选取 n 个元素
- */
-export function pickEleByAlgorithm(
-  array: string[],
-  algorithm: Algorithm,
-  n = 1,
-): string[] {
-  const pickFnMap = {
-    weight: pickWeightEle,
-    random: pickRandomEle,
-  };
-
-  const pickFn = pickFnMap[algorithm];
-
-  if (pickFn) {
-    return pickFn(array, n);
-  }
-
-  /* istanbul ignore next -- @preserve */
-  return pickFnMap[DEFAULT_SURNAME_ALGORITHM](array, n);
-}
-
-/**
- * @private
  *  打乱数组
  */
 export function shuffle(array: string[]): string[] {
@@ -365,81 +346,4 @@ function getGivenNameByGivenNameType(
     }
   }
   return shuffle(result).join('');
-}
-
-/**
- * @private
- * 获取数组范围内随机数
- */
-function randomNumber(a: number, b: number): number {
-  return Math.floor(Math.random() * (b - a + 1));
-}
-
-/**
- * @private
- * 通过算法安全的从数组中随机选取 1 个元素
- */
-function safePickSingleEleByAlgorithm(
-  array: string[],
-  algorithm: Algorithm,
-): string {
-  if (!array || !array.length) {
-    return '';
-  }
-  return pickEleByAlgorithm(array, algorithm, 1)[0];
-}
-
-/**
- * @private
- * 通过随机算法安全的从数组中随机选取 1 个元素
- */
-function pickRandomSingleEle(array: string[]): string {
-  return safePickSingleEleByAlgorithm(array, 'random');
-}
-
-/**
- * @private
- * 通过纯随机算法从数组中随机选取 n 个元素
- */
-function pickRandomEle(array: string[], n = 1): string[] {
-  /* istanbul ignore if -- @preserve */
-  if (!array || !array.length || n <= 0) return [];
-
-  const result = [];
-  const length = array.length;
-  while (result.length < n) {
-    const i = randomNumber(0, length - 1);
-    result.push(array[i]);
-  }
-
-  return result;
-}
-
-/**
- * @private
- * 通过权重算法从数组中随机选取 n 个元素
- */
-function pickWeightEle(array: string[], n = 1): string[] {
-  /* istanbul ignore if -- @preserve */
-  if (!array || !array.length || n <= 0) return [];
-
-  const len = array.length;
-  const totalWeight = (len * (len + 1)) / 2;
-  const result: string[] = [];
-
-  while (result.length < n) {
-    let random = Math.random() * totalWeight;
-    let index = 0;
-    let weight = len;
-
-    while (random >= weight) {
-      random -= weight;
-      index++;
-      weight--;
-    }
-
-    result.push(array[index]);
-  }
-
-  return result;
 }
